Remove unused CatDepthLogin and CategoryDepth imports from multiple files
