+++ ncc_page.py
@@ -7,8 +7,6 @@ from .pages_bots.ncc_page import (
     user_agent,
     NEW_API,
     CatDepth,
-    CatDepthLogin,
-    CategoryDepth,
     change_codes
 )
 
@@ -19,7 +17,5 @@ __all__ = [
     "MainPage",
     "NEW_API",
     "CatDepth",
-    "CatDepthLogin",
-    "CategoryDepth",
     "change_codes",
 ]
